Split oversized indexing/rust.rs along feature boundaries (#509)
`indexing/rust.rs` was 416 lines, over the 400-line budget, mixing
step-attribute detection, function indexing, parameter parsing, and
type rendering. Extract the parameter-parsing helpers
(`parse_function_parameters`, `param_name`, `parameter_is_datatable`,
`parameter_is_step_struct`, `parameter_is_docstring`,
`type_is_string`) into a focused `indexing/rust/params.rs` submodule
with a `//!` header, alongside the existing `type_render` sibling.

`rust.rs` keeps the file-indexing entry points and attribute handling
and drops to 323 lines, so its `scripts/rs-length-allowlist.txt` entry
is removed. No behavioural change; all indexing tests pass unchanged.
