+++ crates/rstest-bdd-server/src/indexing/rust/params.rs
@@ -1,18 +1,54 @@
-//! Parameter classification for indexed Rust step functions.
+//! Function-parameter parsing for Rust step indexing.
 //!
-//! Mirrors the macro behaviour: a data table is expected when a parameter is
-//! named `datatable` or carries `#[datatable]`; a doc string is expected when
-//! a parameter named `docstring` has a `String` type; `#[step_args]` marks a
-//! bundled step-struct parameter.
+//! Converts `syn` function signatures into [`IndexedStepParameter`] records,
+//! classifying each parameter as a data table, doc string, or step-struct
+//! argument using the same conventions as the `rstest-bdd` macros. Type
+//! rendering is delegated to the sibling [`type_render`](super::type_render)
+//! module.
 
-pub(super) fn param_name(pat: &syn::Pat) -> Option<String> {
+use super::IndexedStepParameter;
+use super::type_render;
+
+/// Parse function parameters into indexed step parameters.
+pub(super) fn parse_function_parameters(
+    sig_inputs: &syn::punctuated::Punctuated<syn::FnArg, syn::token::Comma>,
+) -> Vec<IndexedStepParameter> {
+    sig_inputs
+        .iter()
+        .map(|input| match input {
+            syn::FnArg::Receiver(_) => IndexedStepParameter {
+                name: Some("self".to_string()),
+                ty: "Self".to_string(),
+                is_datatable: false,
+                is_docstring: false,
+                is_step_struct: false,
+            },
+            syn::FnArg::Typed(pat_type) => {
+                let name = param_name(&pat_type.pat);
+                let ty = type_render::render_type(&pat_type.ty);
+                let is_datatable = parameter_is_datatable(pat_type, name.as_deref());
+                let is_docstring = parameter_is_docstring(name.as_deref(), &pat_type.ty);
+                let is_step_struct = parameter_is_step_struct(pat_type);
+                IndexedStepParameter {
+                    name,
+                    ty,
+                    is_datatable,
+                    is_docstring,
+                    is_step_struct,
+                }
+            }
+        })
+        .collect()
+}
+
+fn param_name(pat: &syn::Pat) -> Option<String> {
     match pat {
         syn::Pat::Ident(pat_ident) => Some(pat_ident.ident.to_string()),
         _ => None,
     }
 }
 
-pub(super) fn parameter_is_datatable(pat_type: &syn::PatType, name: Option<&str>) -> bool {
+fn parameter_is_datatable(pat_type: &syn::PatType, name: Option<&str>) -> bool {
     if name.is_some_and(|value| value == "datatable") {
         return true;
     }
@@ -29,7 +65,7 @@ pub(super) fn parameter_is_datatable(pat_type: &syn::PatType, name: Option<&str>
 ///
 /// Step struct parameters bundle all placeholders into a single struct,
 /// so they should be counted as step arguments regardless of their name.
-pub(super) fn parameter_is_step_struct(pat_type: &syn::PatType) -> bool {
+fn parameter_is_step_struct(pat_type: &syn::PatType) -> bool {
     pat_type.attrs.iter().any(|attr| {
         attr.path()
             .segments
@@ -38,7 +74,7 @@ pub(super) fn parameter_is_step_struct(pat_type: &syn::PatType) -> bool {
     })
 }
 
-pub(super) fn parameter_is_docstring(name: Option<&str>, ty: &syn::Type) -> bool {
+fn parameter_is_docstring(name: Option<&str>, ty: &syn::Type) -> bool {
     if name.is_none_or(|value| value != "docstring") {
         return false;
     }
@@ -50,78 +86,21 @@ fn type_is_string(ty: &syn::Type) -> bool {
         return false;
     };
 
-    let segments: Vec<&syn::Ident> = type_path.path.segments.iter().map(|s| &s.ident).collect();
-    match segments.as_slice() {
-        [only] => *only == "String",
-        [first, second, third] => {
-            (*first == "std" || *first == "alloc") && *second == "string" && *third == "String"
-        }
-        _ => false,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    //! Unit tests for indexed step-function parameter classification.
-
-    use rstest::rstest;
-
-    use super::{
-        param_name, parameter_is_datatable, parameter_is_docstring, parameter_is_step_struct,
+    let mut segments = type_path.path.segments.iter();
+    let Some(first) = segments.next() else {
+        return false;
     };
-
-    fn typed_parameter(arg: syn::FnArg) -> syn::PatType {
-        match arg {
-            syn::FnArg::Typed(pat_type) => pat_type,
-            syn::FnArg::Receiver(_) => panic!("test parameter should be a typed argument"),
-        }
+    let Some(second) = segments.next() else {
+        return first.ident == "String";
+    };
+    let Some(third) = segments.next() else {
+        return false;
+    };
+    if segments.next().is_some() {
+        return false;
     }
 
-    #[test]
-    fn param_name_ignores_non_identifier_patterns() {
-        let pat_type = typed_parameter(syn::parse_quote!((left, right): (u32, u32)));
-        assert_eq!(param_name(&pat_type.pat), None);
-    }
-
-    #[rstest]
-    #[case::attribute(syn::parse_quote!(#[datatable] rows: Vec<Vec<String>>), true)]
-    #[case::qualified_attribute(
-        syn::parse_quote!(#[rstest_bdd::datatable] rows: Vec<Vec<String>>),
-        true
-    )]
-    #[case::name(syn::parse_quote!(datatable: Vec<Vec<String>>), true)]
-    #[case::plain(syn::parse_quote!(count: u32), false)]
-    fn datatable_parameters_are_detected(#[case] arg: syn::FnArg, #[case] expected: bool) {
-        let pat_type = typed_parameter(arg);
-        let name = param_name(&pat_type.pat);
-        assert_eq!(parameter_is_datatable(&pat_type, name.as_deref()), expected);
-    }
-
-    #[rstest]
-    #[case::attribute(syn::parse_quote!(#[step_args] args: LoginArgs), true)]
-    #[case::qualified_attribute(
-        syn::parse_quote!(#[rstest_bdd::step_args] args: LoginArgs),
-        true
-    )]
-    #[case::plain(syn::parse_quote!(args: LoginArgs), false)]
-    fn step_struct_parameters_are_detected(#[case] arg: syn::FnArg, #[case] expected: bool) {
-        assert_eq!(parameter_is_step_struct(&typed_parameter(arg)), expected);
-    }
-
-    #[rstest]
-    #[case::bare_string(syn::parse_quote!(docstring: String), true)]
-    #[case::std_path(syn::parse_quote!(docstring: std::string::String), true)]
-    #[case::alloc_path(syn::parse_quote!(docstring: alloc::string::String), true)]
-    #[case::wrong_type(syn::parse_quote!(docstring: u32), false)]
-    #[case::wrong_name(syn::parse_quote!(body: String), false)]
-    #[case::reference(syn::parse_quote!(docstring: &str), false)]
-    #[case::two_segments(syn::parse_quote!(docstring: string::String), false)]
-    fn docstring_parameters_require_a_string_type(#[case] arg: syn::FnArg, #[case] expected: bool) {
-        let pat_type = typed_parameter(arg);
-        let name = param_name(&pat_type.pat);
-        assert_eq!(
-            parameter_is_docstring(name.as_deref(), &pat_type.ty),
-            expected
-        );
-    }
+    (first.ident == "std" || first.ident == "alloc")
+        && second.ident == "string"
+        && third.ident == "String"
 }
