+++ scripts/rs-length-allowlist.txt
@@ -7,7 +7,6 @@ crates/rstest-bdd-macros/src/codegen/scenario/runtime/generators/step.rs
 # Added during Phase 7 placeholder count validation feature:
 crates/rstest-bdd-server/src/handlers/diagnostics/compute.rs
 crates/rstest-bdd-server/src/handlers/diagnostics/mod.rs
-crates/rstest-bdd-server/src/indexing/rust.rs
 # Added during tokio-reminders example development (roadmap 9.3.8):
 # Comprehensive doctests for all public methods plus regression test for deferred execution
 examples/tokio-reminders/src/lib.rs
